+++ test/loader.test.js
@@ -48,7 +48,7 @@ describe('loader', () => {
 
   it('should work multiple commonjs exports', async () => {
     const compiler = getCompiler('simple-commonjs2-multiple-export.js', {
-      exposes: ['myOtherGlobal', 'myGlobal.globalObject2 globalObject2'],
+      exposes: ['myOtherGlobal', 'myGlobal.globalObject2|globalObject2'],
     });
     const stats = await compile(compiler);
 
@@ -80,7 +80,7 @@ describe('loader', () => {
 
   it('should work for nested properties for a global object', async () => {
     const compiler = getCompiler('simple-commonjs2-single-export.js', {
-      exposes: ['myGlobal.nested', 'myOtherGlobal.nested foo'],
+      exposes: ['myGlobal.nested', 'myOtherGlobal.nested|foo'],
     });
     const stats = await compile(compiler);
 
@@ -113,9 +113,9 @@ describe('loader', () => {
   it('should work string config', async () => {
     const compiler = getCompiler('simple-module-named-export.js', {
       exposes: [
-        'myGlobal_alias.globalObject6 globalObject6',
-        'myGlobal_alias.globalObject7 globalObject7',
-        'myGlobal_alias.default default',
+        'myGlobal_alias.globalObject6|globalObject6',
+        'myGlobal_alias.globalObject7|globalObject7',
+        'myGlobal_alias.default|default',
         'myGlobal',
         'myOtherGlobal',
       ],
@@ -162,7 +162,7 @@ describe('loader', () => {
           globalName: ['myGlobal_alias', 'globalObject7'],
           localName: 'globalObject7',
         },
-        'myGlobal_alias.default default',
+        'myGlobal_alias.default|default',
       ],
     });
     const stats = await compile(compiler);
@@ -263,8 +263,8 @@ describe('loader', () => {
     );
     const stats = await compile(compiler);
 
-    const webpack4Filename = 'main-b0db002e.bundle.js';
-    const webpack5Filename = 'main-effaed92.bundle.js';
+    const webpack4Filename = 'main-9486a32a.bundle.js';
+    const webpack5Filename = 'main-22966fe7.bundle.js';
     const bundleName =
       webpack.version[0] === '5' ? webpack5Filename : webpack4Filename;
 
@@ -312,7 +312,7 @@ describe('loader', () => {
 
   it('should emit error because of many arguments', async () => {
     const compiler = getCompiler('simple-module-named-export.js', {
-      exposes: ['myGlobal_alias globalObject6 excessArgument'],
+      exposes: ['myGlobal_alias|globalObject6|excessArgument'],
     });
     const stats = await compile(compiler);
